+++ src/app/money/entry.service.ts
@@ -108,7 +108,7 @@ export class EntryService {
     }
 
     saveToStorage(){
-        this.storage.set(this.config.storageKey,JSON.stringify(this.data));
+        // this.storage.set(this.config.storageKey,JSON.stringify(this.data));
     }
 
     newId(){